Fixed AES encryption keysize issue
fastboot_keysize is wrongly set as fastboot_size and this causes
encryption issues for keys other than 128. This is fixed by properly
reading the correct ENV variable on U-boot

Removed U-Boot from the Yocto .wic image as this will be written
separately by BGN-SAT. This is required to not override a secure (or
encrypted) image with plain U-Boot.

Signed-off-by: danie007 <daniel.selvan@jasmin-infotech.com>